# openapi/openapi.yml
openapi: 3.0.0
info:
  title: Dex Demo Embedded REST API
  version: 0.0.1
servers:
  - url: https://testnet.serverfuse.tools/api/v1/
    description: Long-lived testnet.
  - url: https://localnet.dex-demo/api/v1/
    description: URL to the REST API as embedded within the DEX demo client.
...
security:
  - SessionCookie: []
  - CSRFToken: []
paths:
...
  /exchange/orders:
    post:
      summary: Posts an order. Funds will be immediately debited.
      tags:
        - exchange
      security:
        - CSRFToken: []
        - SessionCookie: []
        - OTP: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreationRequest'
      responses:
        '204':
          description: Order successfully posted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderCreationResponse'
    put:
      summary: Modifies an existing order. Funds will immediately debited if the quantity increases, or credited if the quantity decreases.
      tags:
        - exchange
      security:
        - CSRFToken: []
        - SessionCookie: []
        - OTP: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Order successfully modified.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderCreationResponse'
...
  /user/fills:
    get:
      summary: Gets all fills relevant to this user.
      tags:
        - user
      parameters:
        - $ref: '#/components/parameters/StartBlock'
        - $ref: '#/components/parameters/EndBlock'
      responses:
        '200':
          description: The user's fills.
          content:
            application/json:
              schema:
                type: object
                properties:
                  fills:
                    type: array
                    items:
                      $ref: '#/components/schemas/Fill'
...
  /markets/{market_id}/candles:
    get:
      summary: Gets candlestick prices for the provided market.
      tags:
        - market
      parameters:
        - in: path
          name: market_id
          required: true
          description: The ID of the market to fetch candles for.
          schema:
            type: string
        - in: query
          name: start
          description: ISO8601 date of when to start the candle query.
          schema:
            type: string
        - in: query
          name: end
          description: ISO8601 date of when to end the candle query.
          schema:
            type: string
        - in: query
          name: granularity
          description: Granularity of candle data to return.
          schema:
            type: string
            enum: ['1m', '5m', '15m', '60m', '1d']
      responses:
        '200':
          description: Candle data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandlestickResponse'
...
  /markets/{market_id}/book:
    get:
      summary: Gets all open order at latest block by market
      tags:
        - market
      parameters:
        - in: path
          name: market_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The batch state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Spread'
...
  /markets/{market_id}/daily:
    get:
      summary: Daily Price Stats
      tags:
        - market
      parameters:
        - in: path
          name: market_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The stats.
          content:
            application/json:
              schema:
                type: object
                properties:
                  pair:
                    type: string
                  volume:
                    type: string
                  change:
                    type: string
                  last:
                    type: string
                  high:
                    type: string
                  low:
                    type: string
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: uex_session

    CSRFToken:
      type: apiKey
      in: header
      name: X-CSRF-Token

    OTP:
      type: apiKey
      in: header
      name: X-OTP-Token
  parameters:
...
    StartBlock:
      in: query
      name: start_block
      required: false
      description: The block to start counting from.
      schema:
        type: integer
    EndBlock:
      in: query
      name: end_block
      required: false
      description: The block to end counting at.
      schema:
        type: integer
...
  schemas:
    BlockInclusionResponse:
      type: object
      properties:
        block_inclusion:
          type: object
          properties:
            block_number:
              type: string
            transaction_hash:
              type: string
            block_timestamp:
              type: number
...
    OrderCreationRequest:
      type: object
      properties:
        market_id:
          type: string
        direction:
          type: string
          enum: ['BID', 'ASK']
        price:
          type: string
        quantity:
          type: string
        type:
          type: string
          enum: ['MARKET', 'LIMIT']
        time_in_force:
          type: integer
    OrderCreationResponse:
      allOf:
        - $ref: '#/components/schemas/BlockInclusionResponse'
        - $ref: '#/components/schemas/Order'
    Order:
      type: object
      properties:
        id:
          type: string
        market_id:
          type: string
        direction:
          type: string
          enum: ['BID', 'ASK']
        price:
          type: string
        quantity:
          type: string
        status:
          type: string
          enum: ['FILLED', 'CANCELLED', 'OPEN']
        type:
          type: string
          enum: ['MARKET', 'LIMIT']
        time_in_force:
          type: integer
...
    Fill:
      type: object
      allOf:
        - $ref: '#/components/schemas/BlockInclusionResponse'
      properties:
        quantity_filled:
          type: string
        quantity_unfilled:
          type: string
        direction:
          type: string
          enum: ['BID', 'ASK']
        order_id:
          type: string
        pair:
          type: string
        price:
          type: string
        owner:
          type: string
    CandlestickResponse:
      type: object
      properties:
        market_id:
          type: string
        candles:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                description: "ISO8601"
              open:
                type: string
              close:
                type: string
              high:
                type: string
              low:
                type: string
    Spread:
      type: object
      properties:
        block_number:
          type: string
        market_id:
          type: string
        bids:
          type: array
          items:
            $ref: '#/components/schemas/PriceQuantity'
        asks:
          type: array
          items:
            $ref: '#/components/schemas/PriceQuantity'
    PriceQuantity:
      type: object
      properties:
        price:
          type: string
        quantity:
          type: string
...